pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools", "wheel"] # PEP 517

[project]
authors = [
  {name = "Szymon Kozioł", email = "szymonkoz@student.agh.edu.pl"},
  {name = "Inatsi Yermakovich"},
  {name = "Karolina Klimek"},
  {name = "Mateusz Lewandowski"},
  {name = "Patryk Sokoliński"},
  {name = "Błazej Czaicki"},
]
dependencies = [
  "fastapi",
  "uvicorn",
  "pydantic",
  "pydantic[email]",
  "sqlalchemy[asyncio] >= 2.0.13",
  "SQLAlchemy >= 2.0.13",
  "alembic",
  "python-dotenv",
  "psycopg",
  "python-multipart",
  "fastapi-oidc",
  "python-barcode",
  "beautifulsoup4", # Scraping dependencies
  "pyppeteer",
  "pyppeteer_stealth",
  "httpx",
  "python-magic"
]
name = "barcode_api"
readme = "README.md"
version = "0.0.2"

[project.optional-dependencies]
dev = [
  "ruff",
  "black",
  "mypy",
  "pre-commit",
  "pytest",
  "pytest-cov",
  "pytest-mock",
  "pip-tools",
  "sqlalchemy-stubs",
  "types-beautifulsoup4",
  "pytest-asyncio"
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.black]
line-length = 120
target-version = ["py310"]


[tool.setuptools.packages.find]
include = ["barcode_api*"]

[tool.mypy]
exclude = ["build", "dist", "venv", "env", "node_modules"]
